catalog: most of deleteXxxx(int) and delete(Query) methods implemented, the first return the deleted Object and the second will return the numbers of deleted documents

+++ opencga-catalog/src/main/java/org/opencb/opencga/catalog/db/mongodb2/CatalogMongoFileDBAdaptor.java
@@ -146,7 +146,7 @@ public class CatalogMongoFileDBAdaptor extends AbstractCatalogMongoDBAdaptor imp
     }
 
     @Override
-    public QueryResult<Integer> delete(Query query) {
+    public QueryResult<Long> delete(Query query) {
         return null;
     }
